flask_app/requirements.txt
flask>=2.3
flask-cors>=4.0
pillow>=10.0
opencv-python>=4.8
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
catboost>=1.2
scipy>=1.11
werkzeug>=2.3
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.7
